requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cryptography==39.0.0
cycler==0.11.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.38.0
frozendict==2.3.4
frozenlist==1.3.1
future==0.18.2
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
line-bot-sdk==2.3.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.1
mplfinance==0.12.9b1
multidict==6.0.2
multitasking==0.0.11
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
pyimgur==0.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.1
schedule==1.1.0
scipy==1.9.3
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
torch==1.12.1
typing_extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
Werkzeug==2.2.2
wincertstore==0.2
yarl==1.8.1
yfinance==0.1.90
zipp==3.10.0
